Added system-defined UOM Changed to a Long primary key Changed to sequence generator for pk value Made symbols case insensitive Added example to README.md

+++ src/main/java/org/point85/uom/UnitOfMeasure.java
@@ -152,7 +152,7 @@ public class UnitOfMeasure extends Symbolic implements Comparable<UnitOfMeasure>
 	private Integer exponent2;
 
 	// database primary key
-	private Integer primaryKey;
+	private Long primaryKey;
 
 	// optimistic locking version
 	private Integer version;
@@ -185,7 +185,7 @@ public class UnitOfMeasure extends Symbolic implements Comparable<UnitOfMeasure>
 	 * 
 	 * @return Key
 	 */
-	public Integer getKey() {
+	public Long getKey() {
 		return primaryKey;
 	}
 
@@ -195,7 +195,7 @@ public class UnitOfMeasure extends Symbolic implements Comparable<UnitOfMeasure>
 	 * @param key
 	 *            Key
 	 */
-	public void setKey(Integer key) {
+	public void setKey(Long key) {
 		this.primaryKey = key;
 	}
